Added git ignore, split classes into module files
