+++ .gitignore
@@ -1,0 +1,3 @@
+# Git Ignore File
+
+__pycache__/
